Reorganize the project
