+++ mtkcpu/tests/test_branch.py
@@ -1,7 +1,7 @@
 import pytest
 from mtkcpu.utils.common import START_ADDR
-from mtkcpu.tests.utils import MemTestCase, MemTestSourceType, assert_mem_test
-from mtkcpu.tests.models import RegistryContents
+from mtkcpu.utils.tests.utils import MemTestCase, MemTestSourceType, assert_mem_test
+from mtkcpu.utils.tests.registers import RegistryContents
 
 BRANCH_TESTS = [
     MemTestCase(
