+++ mtkcpu/tests/test_upper.py
@@ -1,7 +1,7 @@
 import pytest
 from mtkcpu.utils.common import START_ADDR
-from mtkcpu.tests.utils import MemTestCase, MemTestSourceType, assert_mem_test
-from mtkcpu.tests.models import RegistryContents, MemoryContents
+from mtkcpu.utils.tests.utils import MemTestCase, MemTestSourceType, assert_mem_test
+from mtkcpu.utils.tests.registers import RegistryContents
 from bitstring import Bits
 
 UPPER_TESTS = [
