+++ mtkcpu/utils/tests/exceptions.py
